+++ .gitignore
@@ -6,4 +6,4 @@ node_modules/
 dist/
 .sonarlint/
 package/
-assets/js/**
+web/js/**
