+++ sonar-project.properties
@@ -9,7 +9,7 @@ sonar.projectVersion=1.0
 # If not set, SonarQube starts looking for source code from the directory containing
 # the sonar-project.properties file.
 sonar.sources=.
-sonar.exclusions=assets/,screenshot/,src/js/catalunya-map-path.json
+sonar.exclusions=web/,screenshot/
 
 # Encoding of the source code. Default is default system encoding
 sonar.sourceEncoding=UTF-8
