+++ src/app/catalunya-map-init.js
@@ -1,10 +1,11 @@
-(function(window, catmap) {
-  $(document).ready(function(){
+import catmap from "./catalunya-map";
+
+$(document).ready(function () {
     $.ajax({
         url: window.location.origin + catmap.URL_JSON,
         dataType: 'json',
         async: true,
-        success: function(json) {
+        success: function (json) {
             //Create the map
             var map = catmap.create(catmap.MAP_OPTIONS, json);
             map.loadMapAndText();
@@ -16,5 +17,4 @@
             $("#legend").fadeIn(1000);
         }
     });
-  })
-}(window, window.Catmap));
+})
