+++ src/app/index.js
@@ -1,0 +1,3 @@
+import "./catalunya-map"
+import "./catalunya-map-options-v2"
+import "./catalunya-map-init"
