+++ web/index.html
@@ -16,17 +16,17 @@
   <meta property="og:url"content="http://demo.catalunyamedieval.es/map9/" />
   <meta property="og:title" content="Interactive Map of Catalunya SVG/VML and RaphaëlJs" />
 
-	<script type="text/javascript" src="assets/js/raphael.min.js"></script>
-	<script type="text/javascript" src="assets/js/scale.raphael.min.js"></script>
-	<script type="text/javascript" src="assets/js/jquery-3.2.1.min.js"></script>
-	<script type="text/javascript" src="assets/js/bootstrap.min.js"></script>
+	<script type="text/javascript" src="js/raphael.min.js"></script>
+	<script type="text/javascript" src="js/scale.raphael.min.js"></script>
+	<script type="text/javascript" src="js/jquery-3.2.1.min.js"></script>
+	<script type="text/javascript" src="js/bootstrap.min.js"></script>
 
 	<link href="https://fonts.googleapis.com/css?family=Droid+Sans" rel="stylesheet">
 
-	<link rel="stylesheet" href="assets/css/bootstrap-theme.min.css">
-	<link rel="stylesheet" href="assets/css/bootstrap.min.css">
-	<link rel="stylesheet" href="assets/css/main.css">
-	<link rel="stylesheet" href="assets/css/catalunya-map.min.css">
+	<link rel="stylesheet" href="css/bootstrap-theme.min.css">
+	<link rel="stylesheet" href="css/bootstrap.min.css">
+	<link rel="stylesheet" href="css/main.css">
+	<link rel="stylesheet" href="css/catalunya-map.min.css">
 	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" >
 
 </head>
@@ -158,7 +158,7 @@
 
 </body>
 <footer>
-	<script type="text/javascript" src="assets/js/catalunya-map.min.js"></script>
+	<script type="text/javascript" src="web/js/catalunya-map.min.js"></script>
 </footer>
 
 </html>
